event_metrics.py: Switch to pandas' shape
Use pandas' shape call instead of relying on 'size' to make for a more
intuitive reading experience of the code. In all cases where it was
currently used the number of rows were required but size was used. It
worked because all frames had the same number of columns. This was
probably not obvious for all readers.