event_metrics.py parser: Add 'exclusive-cut'
Introduce an option allowing the user to enable an exclusive cut for
epsilonPID-matrices via the add_isMax_column call.
Alter the plot names of epsilonPID-matrix heatmaps in accordance with
the new exclusive cut. Due to the exclusivity it is not necessary to
adapt the actual value of the cut variable as any value greater than
zero would achieve the same result.