event_metrics.py: Make whiskers optional
Due to the biasing nature of whiskers ('whis') without a physical
motivation it is probable they might at some point they mask the actual
result. Hence remove the default value and ignore them if not explicitly
set.

In the wake of adapting the whiskers, also properly initialize the
selection list ('selection') used for selecting tracks within the
whiskers.