lib.py plot_*: Slight aesthetic adjustments
Specify the x- and y-limits for the plot axes when plotting ROC curves
or other statistics. Both values may never exceed 1 and are always
greater than 0.

Remove the autoformatting of the x axis from the neyman_pearson plot as
there was still plenty of space left and a label tilt was is not
necessary.