event_nn.py: Save PCA per model and reword suffix
Save the PCA transformation per invocation instead of only once. This
makes sense as each dataset has a unique sampling which may result in
different principal components. This bug may cause wrong accuracies
after a later reconstruction. This change fixes the behaviour to at
least make it obvious which PCA should be used for which model.

Furthermore move the savefile_suffix up and make it independent of the
number of layers but add the number of additional layers instead. This
results in new default filenames from now on.